hybris/include: update EGL headers to the latest version
The headers are downloaded from the Khronos website [1] on 2 November
2020.

This is required for the GLVND headers to build properly. I would prefer
to just compile against the system headers. However, the hybris-specific
headers are also in the same directory. I can't make the compiler
include the system header for EGL without either remove our include
path entirely or make the directory a system directory using -isystem
or -idirafter.

[1] https://www.khronos.org/registry/EGL/